+++ src/main/java/org/hermez/reservation/dto/ReservationFormRequest.java
@@ -9,7 +9,7 @@ public class ReservationFormRequest {
   private int courseId;           // 강좌 ID
   private int instructorId;
   private String title;            // 과목명
-  private Double coursePrice;      // 강좌 가격
+  private int coursePrice;      // 강좌 가격
   private String description;      // 강좌 내용
   private String startDate;        // 강좌 시작일
   private String endDate;          // 강좌 종료일
